# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /businesses:
    get:
      operationId: getBusinesses
      summary: List approved businesses (public, with filters)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: string
        - name: lng
          in: query
          schema:
            type: string
        - name: radius
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"

    post:
      operationId: createBusiness
      summary: Create a business listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
...
  /businesses/{id}:
    get:
      operationId: getBusiness
      summary: Get a single business with photos, coupons, and brands
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateBusiness
      summary: Update a business (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

    delete:
      operationId: deleteBusiness
      summary: Delete a business (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /businesses/{id}/photos/{photoId}:
    delete:
      operationId: deletePhoto
      summary: Delete a photo from a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /businesses/{id}/coupons/{couponId}:
    delete:
      operationId: deleteCoupon
      summary: Delete a coupon from a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: couponId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Brand:
      type: object
      required: [id, name, is_featured]
      properties:
        id:
          type: integer
        name:
          type: string
        is_featured:
          type: integer
...
    Photo:
      type: object
      required: [id, business_id, photo_path, display_order]
      properties:
        id:
          type: integer
        business_id:
          type: integer
        photo_path:
          type: string
        display_order:
          type: integer
    Coupon:
      type: object
      required: [id, business_id, image_path]
      properties:
        id:
          type: integer
        business_id:
          type: integer
        image_path:
          type: string
        title:
          type: ["string", "null"]
    Business:
      type: object
      required: [id, owner_id, name, address, status, is_featured, created_at]
      properties:
        id:
          type: integer
        owner_id:
          type: integer
        name:
          type: string
        address:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        hours:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logo_path:
          type: ["string", "null"]
        status:
          type: string
        rejection_reason:
          type: ["string", "null"]
        is_featured:
          type: integer
        created_at:
          type: string
        categories:
          type: array
          items:
            type: string
        brands:
          type: array
          items:
            $ref: "#/components/schemas/Brand"
    BusinessDetail:
      type: object
      required: [id, owner_id, name, address, status, is_featured, created_at, photos, coupons, brands, categories]
      properties:
        id:
          type: integer
        owner_id:
          type: integer
        name:
          type: string
        address:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        hours:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logo_path:
          type: ["string", "null"]
        status:
          type: string
        rejection_reason:
          type: ["string", "null"]
        is_featured:
          type: integer
        created_at:
          type: string
        categories:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        coupons:
          type: array
          items:
            $ref: "#/components/schemas/Coupon"
        brands:
          type: array
          items:
            $ref: "#/components/schemas/Brand"
...
    BusinessInput:
      type: object
      required: [name, address]
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        website:
          type: string
        hours:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            type: string
        brand_ids:
          type: array
          items:
            type: integer
    BusinessUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        website:
          type: string
        hours:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            type: string
        brand_ids:
          type: array
          items:
            type: integer
...