# lib/api-zod/src/generated/types/businessDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { Brand } from './brand';
import type { Coupon } from './coupon';
import type { Photo } from './photo';

export interface BusinessDetail {
  id: number;
  owner_id: number;
  name: string;
  address: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  hours?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logo_path?: string | null;
  status: string;
  /** @nullable */
  rejection_reason?: string | null;
  is_featured: number;
  created_at: string;
  categories: string[];
  photos: Photo[];
  coupons: Coupon[];
  brands: Brand[];
}