# lib/api-zod/src/generated/types/businessInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface BusinessInput {
  name: string;
  address: string;
  phone?: string;
  website?: string;
  hours?: string;
  description?: string;
  categories?: string[];
  brand_ids?: number[];
}