# lib/api-zod/src/generated/types/coupon.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface Coupon {
  id: number;
  business_id: number;
  image_path: string;
  /** @nullable */
  title?: string | null;
}