# lib/api-zod/src/generated/types/photo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface Photo {
  id: number;
  business_id: number;
  photo_path: string;
  display_order: number;
}